refactor: 규칙을 하나로 병합하여 검증 로직을 단순화
