--- src/main/kotlin/omok/domain/rule/OmokGameFinishRule.kt
@@ -1,3 +1,0 @@
-package omok.domain.rule
-
-interface OmokGameFinishRule
