+++ src/test/kotlin/omok/domain/rule/InvalidPositionRuleTest.kt
@@ -23,4 +23,19 @@ class InvalidPositionRuleTest {
         // then
         assertThat(actual).isEqualTo(expected)
     }
+
+    @Test
+    fun `돌이 이미 있는 위치에 돌을 두면 실패한다`() {
+        // given
+        val playingBoard: PlayingBoard = PlayingBoard(rules = OmokRule.rules)
+        val playerStone1: PlayerStone = PlayerStone(StoneColor.BLACK, Position(1, 'A'))
+
+        // when
+        playingBoard.placeStone(playerStone1)
+        val actual = InvalidPositionRule().place(playingBoard.board, playerStone1)
+        val expected = InvalidMove.AlreadyExistStone
+
+        // then
+        assertThat(actual).isEqualTo(expected)
+    }
 }
